+++ src/aks-preview/azext_aks_preview/managed_cluster_decorator.py
@@ -4,6 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 import os
+from base64 import b64encode
 from types import SimpleNamespace
 from typing import Dict, List, Optional, Tuple, TypeVar, Union
 
@@ -41,6 +42,7 @@ from azure.cli.core.azclierror import (
 from azure.cli.core.commands import AzCliCommand
 from azure.cli.core.profiles import ResourceType
 from azure.cli.core.util import get_file_json
+from azure.cli.core.util import read_file_content
 from knack.log import get_logger
 from knack.prompting import prompt_y_n
 
@@ -1298,130 +1300,13 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
                 if disk_driver_version:
                     profile.version = disk_driver_version
             elif disable_disk_driver:
-                msg = "Please make sure there are no existing PVs and PVCs that are used by AzureDisk CSI driver before disabling."
+                msg = (
+                    "Please make sure there are no existing PVs and PVCs "
+                    "that are used by AzureDisk CSI driver before disabling."
+                )
                 if not self.get_yes() and not prompt_y_n(msg, default="n"):
                     raise DecoratorEarlyExitException()
                 profile.enabled = False
-
-        return profile
-
-    def get_file_driver(self) -> Optional[ManagedClusterStorageProfileFileCSIDriver]:
-        """Obtain the value of storage_profile.file_csi_driver
-
-        :return: Optional[ManagedClusterStorageProfileFileCSIDriver]
-        """
-        enable_file_driver = self.raw_param.get("enable_file_driver")
-        disable_file_driver = self.raw_param.get("disable_file_driver")
-
-        if not enable_file_driver and not disable_file_driver:
-            return None
-        profile = self.models.ManagedClusterStorageProfileFileCSIDriver()
-
-        if enable_file_driver and disable_file_driver:
-            raise MutuallyExclusiveArgumentError(
-                "Cannot specify --enable-file-driver and "
-                "--disable-file-driver at the same time."
-            )
-
-        if self.decorator_mode == DecoratorMode.CREATE:
-            if disable_file_driver:
-                profile.enabled = False
-
-        if self.decorator_mode == DecoratorMode.UPDATE:
-            if enable_file_driver:
-                profile.enabled = True
-            elif disable_file_driver:
-                msg = "Please make sure there are no existing PVs and PVCs that are used by AzureFile CSI driver before disabling."
-                if not self.get_yes() and not prompt_y_n(msg, default="n"):
-                    raise DecoratorEarlyExitException()
-                profile.enabled = False
-
-        return profile
-
-    def get_blob_driver(self) -> Optional[ManagedClusterStorageProfileBlobCSIDriver]:
-        """Obtain the value of storage_profile.blob_csi_driver
-
-        :return: Optional[ManagedClusterStorageProfileBlobCSIDriver]
-        """
-        enable_blob_driver = self.raw_param.get("enable_blob_driver")
-        disable_blob_driver = self.raw_param.get("disable_blob_driver")
-
-        if enable_blob_driver is None and disable_blob_driver is None:
-            return None
-
-        profile = self.models.ManagedClusterStorageProfileBlobCSIDriver()
-
-        if enable_blob_driver and disable_blob_driver:
-            raise MutuallyExclusiveArgumentError(
-                "Cannot specify --enable-blob-driver and "
-                "--disable-blob-driver at the same time."
-            )
-
-        if self.decorator_mode == DecoratorMode.CREATE:
-            if enable_blob_driver:
-                profile.enabled = True
-
-        if self.decorator_mode == DecoratorMode.UPDATE:
-            if enable_blob_driver:
-                msg = "Please make sure there is no open-source Blob CSI driver installed before enabling."
-                if not self.get_yes() and not prompt_y_n(msg, default="n"):
-                    raise DecoratorEarlyExitException()
-                profile.enabled = True
-            elif disable_blob_driver:
-                msg = "Please make sure there are no existing PVs and PVCs that are used by Blob CSI driver before disabling."
-                if not self.get_yes() and not prompt_y_n(msg, default="n"):
-                    raise DecoratorEarlyExitException()
-                profile.enabled = False
-
-        return profile
-
-    def get_snapshot_controller(self) -> Optional[ManagedClusterStorageProfileSnapshotController]:
-        """Obtain the value of storage_profile.snapshot_controller
-
-        :return: Optional[ManagedClusterStorageProfileSnapshotController]
-        """
-        enable_snapshot_controller = self.raw_param.get("enable_snapshot_controller")
-        disable_snapshot_controller = self.raw_param.get("disable_snapshot_controller")
-
-        if not enable_snapshot_controller and not disable_snapshot_controller:
-            return None
-
-        profile = self.models.ManagedClusterStorageProfileSnapshotController()
-
-        if enable_snapshot_controller and disable_snapshot_controller:
-            raise MutuallyExclusiveArgumentError(
-                "Cannot specify --enable-snapshot_controller and "
-                "--disable-snapshot_controller at the same time."
-            )
-
-        if self.decorator_mode == DecoratorMode.CREATE:
-            if disable_snapshot_controller:
-                profile.enabled = False
-
-        if self.decorator_mode == DecoratorMode.UPDATE:
-            if enable_snapshot_controller:
-                profile.enabled = True
-            elif disable_snapshot_controller:
-                msg = "Please make sure there are no existing VolumeSnapshots, VolumeSnapshotClasses and VolumeSnapshotContents " \
-                      "that are used by the snapshot controller before disabling."
-                if not self.get_yes() and not prompt_y_n(msg, default="n"):
-                    raise DecoratorEarlyExitException()
-                profile.enabled = False
-
-        return profile
-
-    def get_storage_profile(self) -> Optional[ManagedClusterStorageProfile]:
-        """Obtain the value of storage_profile.
-
-        :return: Optional[ManagedClusterStorageProfile]
-        """
-        profile = self.models.ManagedClusterStorageProfile()
-        if self.mc.storage_profile is not None:
-            profile = self.mc.storage_profile
-        profile.disk_csi_driver = self.get_disk_driver()
-        profile.file_csi_driver = self.get_file_driver()
-        profile.blob_csi_driver = self.get_blob_driver()
-        profile.snapshot_controller = self.get_snapshot_controller()
 
         return profile
 
@@ -1978,6 +1863,31 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         :return: bool
         """
         return self._get_disable_keda(enable_validation=True)
+
+    def get_custom_ca_trust_certificates(self) -> Union[List[bytes], None]:
+        """Obtain the value of custom ca trust certificates.
+
+        :return: List[str] or None
+        """
+        custom_ca_certs_file_path = self.raw_param.get("custom_ca_trust_certificates")
+        if not custom_ca_certs_file_path:
+            return None
+        if not os.path.isfile(custom_ca_certs_file_path):
+            raise InvalidArgumentValueError(
+                "{} is not valid file, or not accessible.".format(
+                    custom_ca_certs_file_path
+                )
+            )
+        # CAs are supposed to be separated with a new line, we filter out empty strings (e.g. some stray new line). We only allow up to 10 CAs
+        file_content = read_file_content(custom_ca_certs_file_path).split(os.linesep + os.linesep)
+        certs = [str.encode(x) for x in file_content if len(x) > 1]
+        if len(certs) > 10:
+            raise InvalidArgumentValueError(
+                "Only up to 10 new-line separated CAs can be passed, got {} instead.".format(
+                    len(certs)
+                )
+            )
+        return certs
 
     def get_defender_config(self) -> Union[ManagedClusterSecurityProfileDefender, None]:
         """Obtain the value of defender.
@@ -2559,6 +2469,22 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
 
         return mc
 
+    def set_up_custom_ca_trust_certificates(self, mc: ManagedCluster) -> ManagedCluster:
+        """Set up Custom CA Trust Certificates for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        ca_certs = self.context.get_custom_ca_trust_certificates()
+        if ca_certs:
+            if mc.security_profile is None:
+                mc.security_profile = self.models.ManagedClusterSecurityProfile()
+
+            mc.security_profile.custom_ca_trust_certificates = ca_certs
+
+        return mc
+
     def set_up_node_restriction(self, mc: ManagedCluster) -> ManagedCluster:
         """Set up security profile nodeRestriction for the ManagedCluster object.
 
@@ -2635,8 +2561,6 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
         mc = self.set_up_image_cleaner(mc)
         # set up cluster snapshot
         mc = self.set_up_creationdata_of_cluster_snapshot(mc)
-        # set up storage profile
-        mc = self.set_up_storage_profile(mc)
         # set up ingress web app routing profile
         mc = self.set_up_ingress_web_app_routing(mc)
         # set up workload auto scaler profile
@@ -2645,6 +2569,8 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
         mc = self.set_up_vpa(mc)
         # set up kube-proxy config
         mc = self.set_up_kube_proxy_config(mc)
+        # set up custom ca trust certificates
+        mc = self.set_up_custom_ca_trust_certificates(mc)
 
         # DO NOT MOVE: keep this at the bottom, restore defaults
         mc = self._restore_defaults_in_mc(mc)
@@ -3069,6 +2995,22 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
 
         return mc
 
+    def update_custom_ca_trust_certificates(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update Custom CA Trust Certificates for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        ca_certs = self.context.get_custom_ca_trust_certificates()
+        if ca_certs:
+            if mc.security_profile is None:
+                mc.security_profile = self.models.ManagedClusterSecurityProfile()
+
+            mc.security_profile.custom_ca_trust_certificates = ca_certs
+
+        return mc
+
     def update_azure_monitor_profile(self, mc: ManagedCluster) -> ManagedCluster:
         """Update azure monitor profile for the ManagedCluster object.
         :return: the ManagedCluster object
@@ -3230,8 +3172,6 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc = self.update_node_restriction(mc)
         # update image cleaner
         mc = self.update_image_cleaner(mc)
-        # update stroage profile
-        mc = self.update_storage_profile(mc)
         # update workload auto scaler profile
         mc = self.update_workload_auto_scaler_profile(mc)
         # update azure monitor metrics profile
@@ -3246,5 +3186,7 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc = self.update_outbound_type_in_network_profile(mc)
         # update kube proxy config
         mc = self.update_kube_proxy_config(mc)
+        # update custom ca trust certificates
+        mc = self.update_custom_ca_trust_certificates(mc)
 
         return mc